TransactionService/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
azure-core==1.35.1
azure-identity==1.25.0
azure-keyvault-secrets==4.10.0
azure-servicebus==7.14.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
fastapi==0.116.1
frozenlist==1.7.0
h11==0.16.0
httptools==0.6.4
idna==3.10
isodate==0.7.2
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.4
propcache==0.3.2
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
